--- simple_classifier/Classifiers/factory/ClassifierFactory.py
@@ -1,5 +1,0 @@
-from simple_classifier.utils.BaseFactory import BaseFactory
-
-
-class ClassifierFactory(BaseFactory):
-    pass
