--- simple_classifier/SplitterDataset/factory/SplitterDatasetFactory.py
@@ -1,5 +1,0 @@
-from simple_classifier.utils.BaseFactory import BaseFactory
-
-
-class SplitterDatasetFactory(BaseFactory):
-    pass
