+++ src/Classifier/factory/ClassifierFactory.py
@@ -1,0 +1,5 @@
+from src.Utilities.BaseFactory import BaseFactory
+
+
+class ClassifierFactory(BaseFactory):
+    pass
