+++ src/SplitterDataset/factory/SplitterDatasetFactory.py
@@ -1,0 +1,5 @@
+from src.Utilities.BaseFactory import BaseFactory
+
+
+class SplitterDatasetFactory(BaseFactory):
+    pass
